reorganizing unit tests
