+++ src/components/page-header/page-header.jsx
@@ -3,9 +3,9 @@ import PropTypes from "prop-types";
 import { PageHeaderWrapper } from "./page-header.styles";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-const PageHeader = ({text, icon, textCenter, fixedPosition}) => {
+const PageHeader = ({text, icon, textCenter, fixedPosition, dataTestId}) => {
   return (
-    <PageHeaderWrapper textCenter={textCenter} fixedPosition={fixedPosition}>
+    <PageHeaderWrapper data-testid={dataTestId} textCenter={textCenter} fixedPosition={fixedPosition}>
       <h1>
         {icon && <FontAwesomeIcon data-testid="pageHeaderIcon" icon={icon} fixedWidth/>}
         {text}
@@ -18,7 +18,8 @@ PageHeader.propTypes = {
   text: PropTypes.string.isRequired,
   icon: PropTypes.object,
   textCenter: PropTypes.bool,
-  fixedPosition: PropTypes.bool
+  fixedPosition: PropTypes.bool,
+  dataTestId: PropTypes.string
 };
 
 export default PageHeader;
