+++ src/store/reducers/root.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 import { connectRouter } from 'connected-react-router'
-import sidebar from "./sidebar";
+import sidebar from "./sidebar/sidebar";
 
 const createRootReducer = (history) => combineReducers({
   router: connectRouter(history),
